move network views to separate file
